New Loading Screens
So I was tired of looking at these WW2 screens. And I made new ones. The
art style is a bit different on some of them, things can be changed
around and new ones added But I think it looks good! All source imagery
is in the public domain and can be used without reference. Check it out!