BRA oob and county file fixes and adjustments
Changed up a few techs, such as AMX no longer being a CAS but a strike fighter and techs for nuclear sub being built, as well as Tamandaré Corvettes. WIll follow this up with name changes and after that the new icons. I was planning on sending them all together, but after recently deleting all my work accidentally I decided to send it in 3 parts.