Fixed some cores/claims to be work with new system
Don't ask why this is on this branch...

Taiwan gets claims not cores on Mongolia and Arunachal Pradesh.
China gets a claim, not core on Arunachal Pradesh.
Tibet gets claims not cores on Qinghai and Kham.

Tibet SAR focus nolonger uncores Qinghai and Kham, and Aksai Chin (which is part of Xinjiang province) remains part of China.
Gave Tibet SAR flag and more "varied" politicts.